+++ post_install_arch/scripts/yay.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 
+if [ -x "$(command -v yay)" ]; then
+  echo "Yay has already been installed. Skipping."
+  exit 0;
+fi
+
 sudo pacman -Syu
+
+cd $HOME
 
 # Just to be sure
 rm -rf yay
